gradio_app/pyproject.toml
[project]
name = "gradio-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio==5.14.0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b36",
    "azure-search-documents>=11.6.0b4",
    "semantic-kernel[azure]>=1.31.0",
    "uvicorn>=0.34.2",
    "starlette>=0.46.2",
    "authlib>=1.6.0",
    "fastapi>=0.115.12",
    "itsdangerous>=2.2.0",
    "msal>=1.30.0",
    "azure-identity>=1.19.0",
    "PyJWT>=2.8.0",
    "python-multipart>=0.0.6",
    "starlette-session>=0.4.3",
]
